requirements.txt
APScheduler==3.10.1
Flask==2.3.2
Flask-APScheduler==1.13.1
Flask-CORS==3.0.10
Flask-GraphQL==2.0.1
Flask-Login==0.6.2
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.4
Flask-WTF==1.1.1
graphene==2.1.8
gunicorn==21.1.0
numpy==1.26.4
pandas==2.0.3
psycopg2-binary==2.9.6
python-dotenv==1.0.0
requests==2.31.0
Werkzeug==2.3.6
email-validator==2.0.0